# Import more fields for attendance events 