--- a/app/controllers/schools_controller.rb
+++ b/app/controllers/schools_controller.rb
@@ -173,7 +173,7 @@ class SchoolsController < ApplicationController
       last_name: student.last_name,
       id: student.id,
       homeroom_label: homeroom_label(student.homeroom),
-      absences: student.dashboard_absences,
+      absences: student.dashboard_absences.as_json(only: [:student_id, :occurred_at]),
       sst_notes: student.sst_notes,
     }
   end
@@ -184,7 +184,7 @@ class SchoolsController < ApplicationController
       last_name: student.last_name,
       id: student.id,
       homeroom_label: homeroom_label(student.homeroom),
-      tardies: student.dashboard_tardies,
+      tardies: student.dashboard_tardies.as_json(only: [:student_id, :occurred_at]),
       sst_notes: student.sst_notes,
     }
   end