--- a/app/importers/rows/attendance_row.rb
+++ b/app/importers/rows/attendance_row.rb
@@ -16,6 +16,8 @@ class AttendanceRow < Struct.new(:row)
       def valid?; false end
 
       def errors; NullEventErrors.new end
+
+      def assign_attributes(attributes); end
     end
 
     def self.find_or_initialize_by(_)
@@ -31,6 +33,13 @@ class AttendanceRow < Struct.new(:row)
     attendance_event = attendance_event_class.find_or_initialize_by(
       occurred_at: row[:event_date],
       student_id: student.try(:id),
+    )
+
+    attendance_event.assign_attributes(
+      dismissed: row[:dismissed],
+      excused: row[:excused],
+      reason: row[:reason],
+      comment: row[:comment],
     )
 
     return attendance_event