--- a/app/models/student.rb
+++ b/app/models/student.rb
@@ -21,10 +21,10 @@ class Student < ActiveRecord::Base
   has_many :sections, through: :student_section_assignments
 
   has_many :dashboard_tardies, -> {
-    where('occurred_at >= ?', 1.year.ago).select('student_id', 'occurred_at')
+    where('occurred_at >= ?', 1.year.ago)
   }, class_name: "Tardy"
   has_many :dashboard_absences, -> {
-    where('occurred_at >= ?', 1.year.ago).select('student_id', 'occurred_at')
+    where('occurred_at >= ?', 1.year.ago)
   }, class_name: "Absence"
 
   has_many :sst_notes, -> { where(event_note_type_id: 300) }, class_name: "EventNote"