--- a/db/migrate/20180426135126_add_new_fields_to_attendance_events.rb
+++ b/db/migrate/20180426135126_add_new_fields_to_attendance_events.rb
@@ -1,0 +1,13 @@
+class AddNewFieldsToAttendanceEvents < ActiveRecord::Migration[5.1]
+  def change
+    add_column :absences, :dismissed, :boolean
+    add_column :absences, :excused, :boolean
+    add_column :absences, :reason, :string
+    add_column :absences, :comment, :string
+
+    add_column :tardies, :dismissed, :boolean
+    add_column :tardies, :excused, :boolean
+    add_column :tardies, :reason, :string
+    add_column :tardies, :comment, :string
+  end
+end