--- a/spec/importers/file_importers/attendance_importer_spec.rb
+++ b/spec/importers/file_importers/attendance_importer_spec.rb
@@ -24,18 +24,28 @@ RSpec.describe AttendanceImporter do
         let!(:student) { FactoryBot.create(:student, local_id: '1') }
         let(:date) { '2005-09-16' }
 
-        context 'row with absence' do
+        context 'row with absence (district has dismissed/reason/excused/comment fields)' do
           let(:row) {
-            { event_date: date, local_id: '1', absence: '1', tardy: '0' }
+            { event_date: date, local_id: '1', absence: '1', tardy: '0',
+              dismissed: '0', reason: 'Medical', excused: '0', comment: 'Received doctor note.' }
           }
 
           it 'creates an absence' do
-
             expect {
               attendance_importer.import_row(row)
             }.to change {
               Absence.count
             }.by 1
+          end
+
+          it 'sets the right attributes' do
+            attendance_importer.import_row(row)
+            absence = Absence.first
+
+            expect(absence.dismissed).to eq false
+            expect(absence.reason).to eq 'Medical'
+            expect(absence.excused).to eq false
+            expect(absence.comment).to eq 'Received doctor note.'
           end
 
           it 'creates only 1 absence if run twice' do
@@ -60,6 +70,33 @@ RSpec.describe AttendanceImporter do
               student.tardies.size
             }.by 0
           end
+        end
+      end
+
+      context 'row with absence (district does not have dismissed/reason/excused/comment fields)' do
+        let!(:student) { FactoryBot.create(:student, local_id: '1') }
+        let(:date) { '2005-09-16' }
+
+        let(:row) {
+          { event_date: date, local_id: '1', absence: '1', tardy: '0' }
+        }
+
+        it 'creates an absence' do
+          expect {
+            attendance_importer.import_row(row)
+          }.to change {
+            Absence.count
+          }.by 1
+        end
+
+        it 'sets the right attributes' do
+          attendance_importer.import_row(row)
+          absence = Absence.first
+
+          expect(absence.dismissed).to eq nil
+          expect(absence.reason).to eq nil
+          expect(absence.excused).to eq nil
+          expect(absence.comment).to eq nil
         end
       end