… spec/fixtures/fake_attendance_export.txt