# Who is this PR for?

Educators who rely on Student Insights for analysis and visualizations about student attendance. 

# What problem does this PR fix?

Some attendance events are excused; Insights doesn't surface this right now, which makes educators need to turn to Aspen for more detailed data.

# What does this PR do?

Builds on #1650 by pulling in the new columns to our database. 

I tested what this would look like in production by running an import process into my local machine using this branch. Here's what I found (summary table above, calculations below):

### Summary table

field | Absences | Tardies
---|---|---
dismissed == true | 0.66% | 1.97%
excused == true | 6.27% | 2.78%
reason present? | 29.6% | 10.0%
comment present? | 18.3% | 4.4% 

### Exact queries + results

```
# Overall counts
2.3.0 :002 > Absence.count
 => 226182
2.3.0 :003 > Tardy.count
 => 188479

# Absences: dismissed 
2.3.0 :005 > Absence.where(dismissed: true).count
 => 1492
fraction of total: 0.006596457719889292 => 0.66%
2.3.0 :006 > Absence.where(dismissed: false).count
 => 224690
fraction of total: 0.9934035422801107 => 99.34%

# Absences: excused 
2.3.0 :007 > Absence.where(excused: true).count
 => 14185
fraction of total: 0.06271498174036838 => 6.27%
2.3.0 :008 > Absence.where(excused: false).count
 => 211997
fraction of total: 0.9372850182596316 => 93.73%

# Absences: reason
2.3.0 :049 > Absence.where(reason: ["", nil]).count
 => 159325
 fraction of total: 0.7044106073869716 => 70.4%
2.3.0 :050 > Absence.where.not(reason: ["", nil]).count
 => 66857
 fraction of total: 0.29558939261302847 => 29.6%

# Absences: comment
2.3.0 :052 > Absence.where(comment: ["", nil]).count
 => 184833
 fraction of total: 0.8171870440618617 => 81.7%
2.3.0 :053 > Absence.where.not(comment: ["", nil]).count
 => 41349
 fraction of total: 0.18281295593813832 => 18.3%

# Tardies: dismissed
2.3.0 :012 > Tardy.where(dismissed: true).count
 => 3713
fraction of total: 0.019699807405599564 => 1.97%
2.3.0 :013 > Tardy.where(dismissed: false).count
 => 184766
fraction of total: 0.9803001925944005 => 98.03%

# Tardies: excused
2.3.0 :015 > Tardy.where(excused: true).count
 => 5240
fraction of total: 0.027801505738039783 => 2.78%
2.3.0 :016 > Tardy.where(excused: false).count
 => 183239
fraction of total: 0.9721984942619603 => 97.2%

# Tardies: reason
2.3.0 :062 > Tardy.where(reason: ["", nil]).count
 => 169716
fraction of total: 0.9004504480605267 => 90.0%
2.3.0 :063 > Tardy.where.not(reason: ["", nil]).count
 => 18763
fraction of total: 0.09954955193947336 => 10.0%

# Tardies: comment
2.3.0 :064 > Tardy.where(comment: ["", nil]).count
 => 180251
fraction of total: 0.9563452692342383 => 95.6%
2.3.0 :065 > Tardy.where.not(comment: ["", nil]).count
 => 8228
fraction of total: 0.0436547307657617 => 4.4%
```

# Question

One thing I don't know: what does "dismissed" mean in this context? @kevinrobinson @edavidsonsawyer do you know this one? If not, maybe we can ask Uri tomorrow. Since it's the least frequent of all the new columns, I'm wondering if it's used by all Somerville schools. Or maybe there's another reason it has such a low frequency. 

# "Reason" categories 

There are a fixed number of categories used in the "reason" column for absences and tardies, in contrast with "comment" which is free-form.

```
=> puts Absence.pluck(:reason).uniq.compact.sort

Appointment
Bereavement
Holiday
Late
Legal
Medical
No Reason
Nurse
Parent
School
Sick
Suspended
Transportation
Truant
Vacation
Walked Out
Weather
Withdrawal
```