@alexsoble 2018-04-26
@kevinrobinson Updated to use `PerDistrict` in the import code / specs