@alexsoble 2018-04-27
In a5bedf6332cd1faed27f4d63be513f39976cc303 I limited the fields we're sending down the Absences and Tardies dashboard as per @kevinrobinson's sharp suggestion here: https://github.com/studentinsights/studentinsights/pull/1653#discussion_r184662534

I did some measuring in my local environment for the queries loading the Absences dashboard with `ENV['MORE_DEMO_STUDENTS'] = true`. Here's the impact I found on server side performance. There are probably also positive impacts on client side performance here too but I didn't measure them.

**Before using `select` to limit fields**:

```
[0] Completed 200 OK in 1117ms (Views: 577.5ms | ActiveRecord: 54.1ms)
[0] Completed 200 OK in 1192ms (Views: 593.5ms | ActiveRecord: 67.0ms)
[0] Completed 200 OK in 1237ms (Views: 591.4ms | ActiveRecord: 47.6ms)
[0] Completed 200 OK in 1113ms (Views: 590.8ms | ActiveRecord: 46.7ms)
[0] Completed 200 OK in 1068ms (Views: 568.1ms | ActiveRecord: 42.3ms)

Average load time: 1145ms
```

**After using `select` to limit fields**:

```
[0] Completed 200 OK in 1384ms (Views: 471.7ms | ActiveRecord: 88.7ms)
[0] Completed 200 OK in 981ms (Views: 391.0ms | ActiveRecord: 43.3ms)
[0] Completed 200 OK in 941ms (Views: 397.3ms | ActiveRecord: 38.7ms)
[0] Completed 200 OK in 1288ms (Views: 520.5ms | ActiveRecord: 42.8ms)
[0] Completed 200 OK in 847ms (Views: 374.6ms | ActiveRecord: 41.3ms)

Average load time: 1088ms
```

This shows a ~5% improvement in the server side performance from limiting the fields we're sending down; it seems reasonable to expect a similar impact on the client side performance too.