@alexsoble 2018-04-27
@edavidsonsawyer Take a look at the change I made to `dashboard_tardies` and `dashboard_absences` in a5bedf6332cd1faed27f4d63be513f39976cc303, these limit the columns we're sending down to just `student_id` and `occurred_at`. 

This means `created_at` and `updated_at` will be filtered out since we don't use them in the Dashboard view, because they're referring to when the records were created as opposed to when the attendance event occurred.

My assumption was that this would have no impact on the Absence/Tardies dashboard functionality besides speeding it up a bit, and that seemed to be the case when I tested it out locally. Wanted to make sure you got a chance to take a  👀 since you've written so much of the Absence/Tardy dashboard code.

Also @edavidsonsawyer, after this PR merges, you can run`rake db:reset` to re-seed the data and get fake but representative values for the new attendance event columns from 1fd830122627d2b30fb936a070df59434538c8db.