@alexsoble 2018-04-27
@kevinrobinson Hmmmm, `select` isn't causing any bugs in this PR and is improving performance, so I think it should be fine to keep it here. We're using `select` in `StudentsQueryHelper` right now and it hasn't introduced any bugs.