@kevinrobinson 2018-04-27
@alexsoble 👍 cool it's just my Rails brain that's buggy then :)