@kevinrobinson 2018-04-27
@alexsoble Awesome!

Yeah so my understanding of `select` is that it applies at the level of the AR query and model.  It adds to the "SELECT" clause in the SQL that queries the database, and says "only get me the student_id and occurred_at field from the database."  So then those AR models that you get back literally only have those fields in them.  They don't have id fields or anything else like reason, etc. because `select` tells AR not to fetch them from the database.  I found that led to me making a few different mistakes, because most model methods won't work as you expect then (eg, other associations won't work, methods that read other fields on the model will raise).

It does end up influencing what happens when you later call `as_json`, since those fields aren't on the model they aren't serialized.  But if you want to limit the serialization then the more direct way is to pass `:only` when calling `as_json`.  So that's what I ended up doing, since reducing the bytes sent to the client reduced response time more than reducing the fields read from the db (since the db still has to read all the same rows anyway).  So one of my takeaways from that was `as_json(:only => ... )` was a simpler way to optimize the bytes sent to the client, and that might be a better place to start than bytes read from the db.

I was thinking of this more as "limit the data sent to the browser" than "limit the data read from the db."  But it might be different here, or you might have found other approaches that make more sense for this situation!