@alexsoble 2018-04-27
@kevinrobinson, Alrighty subbed in `as_json` instead of `select` in 1ed017825cf253194ef1d62a3c16e6f6859eb0ae. I reviewed the absences and tardies pages manually to confirm that they are behaving just like before. @kevinrobinson, can you do the same QA and then flash a 👍 if this PR seems good to merge and deploy?